src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bentoml==1.0.23
build==0.10.0
category-encoders==2.6.1
cattrs==23.1.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
circus==0.18.0
click==8.1.4
click-option-group==0.5.6
cloudpickle==2.2.1
comm==0.1.3
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
deepmerge==1.1.0
Deprecated==1.2.14
deprecation==2.1.0
deptry==0.12.0
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.40.0
frozenlist==1.3.3
fs==2.4.16
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.0.1
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaleido==0.2.1
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.4.2
multidict==6.0.4
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aiohttp-client==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
orjson==3.9.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip-requirements-parser==32.0.1
pip-tools==6.14.0
platformdirs==3.8.1
plotly==5.15.0
plotly-resampler==0.9.0
pmdarima==2.0.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pycaret==3.0.4
Pygments==2.15.1
PyJWT==2.7.0
pynvml==11.5.0
pyod==1.1.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
querystring-parser==1.2.4
referencing==0.29.1
requests==2.31.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.8.10
schema==0.7.5
schemdraw==0.15
scikit-base==0.5.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.11.1
simple-di==0.1.5
six==1.16.0
sktime==0.20.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.28.0
statsmodels==0.14.0
tabulate==0.9.0
tbats==1.1.3
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.9.0
tsdownsample==0.1.2
typing_extensions==4.7.1
Unidecode==1.3.6
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.15.0
wurlitzer==3.0.3
xxhash==3.2.0
yarl==1.9.2
yellowbrick==1.5
zipp==3.16.0
